Polish the noise up a little.

+++ src/main/java/vini2003/xyz/eco/common/util/NoiseUtils.java
@@ -13,7 +13,7 @@ public class NoiseUtils {
 		return NormalizeUtils.normalize(sample, -1.0F, 1.0F, 0.0F, 1.0F);
 	}
 	
-	public static float getNoise(FastNoiseLite noise, int x, int z, int octaves) {
+	public static float getNoise(FastNoiseLite noise, float x, float z, int octaves) {
 		float total = 0.0F;
 		
 		float frequency = 1.0F;
